Services and (simple Repo) implemented

+++ MeasurementUnitsManagement.Models/Entities/UnitFactorModel.cs
@@ -2,9 +2,10 @@ namespace MeasurementUnitsManagement.Models.Entities
 {
     public class UnitFactorModel : BaseUnit
     {
-        public decimal RatioToTheBaseUnit { get; set; }
+        public double RatioToTheBaseUnit { get; set; }
 
         public int BaseUnitId { get; set; }
-        public BaseUnitModel BaseUnitInfo { get; set; }
+        public BaseUnitModel BaseUnitInfo { get; set; } 
+        
     }
 }
